requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
groq
google-generativeai
pymilvus
numpy
pdfplumber
python-docx
python-pptx
pandas
openpyxl
beautifulsoup4
lxml
streamlit
requests
preocr
mcp
rich
